requirements.txt
Django==5.2.6
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.32.5
gunicorn==21.2.0
whitenoise==6.6.0
